icon-theme-spec: Add icon scale support
This adds support for "pre-scaled" icons aimed at higher resolution displays.
For instance, an icon at scale 2 and size 48 (48@2x) would have the
same kind of detail as a 48@1x icon, but a pixel size of 96x96. A
a 96@1x icon would also be 96x96 but would have more details that would
be hard to see on a high resolution screen.